@@ -1,6 +1,6 @@
 import {RoundResults} from '@doko/common';
 import {useCallback} from 'react';
-import {useHistory} from 'react-router-dom';
+import {useNavigate} from 'react-router-dom';
 import {useSortedGames} from '../Games';
 import {usePlayersWithStats} from '../Players';
 import {usePatchRound, useRound} from '../Rounds';
@@ -10,7 +10,7 @@ export function useFinishRound() {
   const patchRound = usePatchRound();
   const playersWithStats = usePlayersWithStats(true);
   const sortedGames = useSortedGames();
-  const history = useHistory();
+  const navigate = useNavigate();
   const lastGame = sortedGames[sortedGames.length - 1];
   return useCallback((forcePrematureEnd = false) => {
     if (!round || !lastGame || (!lastGame.data.isLastGame && !forcePrematureEnd)) {
@@ -28,6 +28,6 @@ export function useFinishRound() {
       endDate: Math.round(Date.now() / 1000),
       data: {results},
     });
-    history.push(`/group/${round.groupId}/rounds`);
-  }, [history, lastGame, patchRound, playersWithStats, round, sortedGames.length]);
+    navigate(`/group/${round.groupId}/rounds`);
+  }, [navigate, lastGame, patchRound, playersWithStats, round, sortedGames.length]);
 }
