@@ -1,7 +1,7 @@
 import {RoundsRemove} from '@doko/common';
 import {useCallback} from 'react';
 import {useDispatch} from 'react-redux';
-import {useHistory} from 'react-router-dom';
+import {useNavigate} from 'react-router-dom';
 import {useSortedGames} from '../Games';
 import {LoguxDispatch} from '../Logux';
 import {useRound} from '../Rounds';
@@ -9,7 +9,7 @@ import {useRound} from '../Rounds';
 export function useRemoveRound() {
   const round = useRound();
   const currentGames = useSortedGames();
-  const history = useHistory();
+  const navigate = useNavigate();
   const dispatch = useDispatch<LoguxDispatch>();
   return useCallback((): void => {
     if (!round) {
@@ -19,6 +19,6 @@ export function useRemoveRound() {
       return;
     }
     dispatch.sync<RoundsRemove>({id: round.id, groupId: round.groupId, type: 'rounds/remove'});
-    history.push(`/group/${round.groupId}`);
-  }, [round, currentGames.length, dispatch, history]);
+    navigate(`/group/${round.groupId}`);
+  }, [round, currentGames.length, dispatch, navigate]);
 }
