@@ -1,5 +1,4 @@
 import React, {ReactElement} from 'react';
-import {useRouteMatch} from 'react-router-dom';
 import {Divider} from 'semantic-ui-react';
 import {asLink} from '../../AsLink';
 import {useGroup} from '../../store/Groups';
@@ -7,12 +6,11 @@ import FullPot from '../pot/FullPot';
 
 export default function Pot(): ReactElement | null {
   const group = useGroup();
-  const {url} = useRouteMatch();
   if (!group?.settings.eurosPerPointDiffToTopPlayer) {
     return null;
   }
   return <>
-    <FullPot as={asLink(`${url}/pot`)}/>
+    <FullPot as={asLink(`pot`)}/>
     <Divider/>
   </>;
 }
