@@ -1,20 +1,18 @@
 import dayjs from 'dayjs';
 import React, {ReactElement} from 'react';
-import {useRouteMatch} from 'react-router-dom';
 import {Icon, Label} from 'semantic-ui-react';
 import {asLink} from '../../AsLink';
 import {useAddRound} from '../../store/Round/AddRound';
 import {useSortedRounds} from '../../store/Rounds';
 
 export default function RoundsInfo(): ReactElement {
-  const {url} = useRouteMatch();
   const rounds = useSortedRounds();
   const addRound = useAddRound();
   const lastRound = rounds[0];
 
   return <section>
     <div className='memberDetail'>
-      <Label as={asLink(`${url}/rounds`)} color={'orange'}>
+      <Label as={asLink(`rounds`)} color={'orange'}>
         Alle Runden
         <Label.Detail>
           {rounds.length} <Icon name={'bullseye'}/>
@@ -23,7 +21,7 @@ export default function RoundsInfo(): ReactElement {
     </div>
 
     {!!lastRound && <div className='memberDetail'>
-      <Label as={asLink(`${url}/rounds/round/${lastRound.id}`)} color={'blue'}>
+      <Label as={asLink(`rounds/round/${lastRound.id}`)} color={'blue'}>
         {lastRound.endDate ? 'Letzte' : 'Aktuelle'} Runde
         <Label.Detail>
           {dayjs.unix(lastRound.startDate).format('DD.MM.YYYY HH:mm')} <Icon name={'clock'}/>
