@@ -1,6 +1,7 @@
 import {bockGamesTranslations, bockInBockBehaviorOptions, extraPointsTranslations} from '@doko/common';
 import React, {ReactElement} from 'react';
 import {Divider, Form, Header} from 'semantic-ui-react';
+import Page from '../../Page';
 import {useGroup} from '../../store/Groups';
 import AllowedSoloTypes from './settings/AllowedSoloTypes';
 import BockEffect from './settings/BockEffect';
@@ -12,25 +13,27 @@ import RadioGroup from './settings/RadioGroup';
 export default function Settings(): ReactElement | null {
   const {settings} = useGroup()!;
   const hasBock = Object.values(settings.bockGames).some((val) => val);
-  return <section>
-    <Form>
-      <Name/>
-      <Checkbox path={`dynamicRoundDuration`} label={'Dynamische Rundenlänge - sonst 24 Spiele'}/>
-      <AllowedSoloTypes/>
-      <CheckboxGroup label={'Zusatzpunkte bei ...'} options={extraPointsTranslations} parentKey={'extraPoints'}/>
+  return <Page displayName={'Einstellungen'}>
+    <section>
+      <Form>
+        <Name/>
+        <Checkbox path={`dynamicRoundDuration`} label={'Dynamische Rundenlänge - sonst 24 Spiele'}/>
+        <AllowedSoloTypes/>
+        <CheckboxGroup label={'Zusatzpunkte bei ...'} options={extraPointsTranslations} parentKey={'extraPoints'}/>
 
-      <Divider section/>
+        <Divider section/>
 
-      <section>
-        <Header as='h4'>Bock</Header>
-        <CheckboxGroup label={'Bockspiele nach ...'} options={bockGamesTranslations} parentKey={'bockGames'}/>
-        {hasBock && <>
-          <RadioGroup label={'Bei Bock im Bock ...'}
-                      options={bockInBockBehaviorOptions}
-                      parentKey={'bockInBockBehavior'}/>
-          <BockEffect/>
-        </>}
-      </section>
-    </Form>
-  </section>;
+        <section>
+          <Header as='h4'>Bock</Header>
+          <CheckboxGroup label={'Bockspiele nach ...'} options={bockGamesTranslations} parentKey={'bockGames'}/>
+          {hasBock && <>
+            <RadioGroup label={'Bei Bock im Bock ...'}
+                        options={bockInBockBehaviorOptions}
+                        parentKey={'bockInBockBehavior'}/>
+            <BockEffect/>
+          </>}
+        </section>
+      </Form>
+    </section>
+  </Page>;
 }
