@@ -1,5 +1,4 @@
 import React, {ReactElement} from 'react';
-import {useRouteMatch} from 'react-router-dom';
 import {Header, Icon, Label} from 'semantic-ui-react';
 import {asLink} from '../../AsLink';
 import {useSortedGames} from '../../store/Games';
@@ -7,7 +6,6 @@ import AddGame from './AddGame';
 import RoundEndInfo from './RoundEndInfo';
 
 export default function GamesInfo(): ReactElement {
-  const {url} = useRouteMatch();
   const sortedGames = useSortedGames();
   const lastGame = sortedGames[sortedGames.length - 1];
   const hasOpenGame = !!lastGame && !lastGame.data.isComplete;
@@ -16,7 +14,7 @@ export default function GamesInfo(): ReactElement {
     <Header as='h4'>Spiele</Header>
 
     <div className='memberDetail'>
-      <Label as={asLink(`${url}/games`)} color={'orange'}>
+      <Label as={asLink(`games`)} color={'orange'}>
         Alle Spiele
         <Label.Detail>
           {sortedGames.length} <Icon name={'hashtag'}/>
@@ -27,7 +25,7 @@ export default function GamesInfo(): ReactElement {
     <RoundEndInfo/>
 
     {hasOpenGame && <div className='memberDetail'>
-      <Label as={asLink(`${url}/games/game/${lastGame.id}`)} color={'blue'}>
+      <Label as={asLink(`games/game/${lastGame.id}`)} color={'blue'}>
         Laufendes Spiel
       </Label>
     </div>}
